add simplistic web browser support under certain circumstances
GET of a session with Accept: text/html while no session is active,
the "session_uri" config parameter is set, configured clients provider
supports login interface, and ?referrer=URI query parameter may be
optionally set:

  will respond with HTML page containing basic login form, including
  hidden referrer field containing query parameter.

POST for a session with referrer input, on successful login, will
issue a 303 See Other redirect to the referrer.  Otherwise, POST
returns basic session key as before.

Thus, an application MAY support HTML content-negotiated login for its
REST URIs if it wants to, threading through this basic form rather
than requiring a full AJAX UI.  But it is up to application whether to
do this in general or not.