Database providers now use libpq DSN connection strings.
All of the providers that were using database backed storage 
have been updated to use the new format. By default a string
of dsn='dbname=' is equivalent to the former argument of
dbname=''. It will connect over domain socket to the data-
base named the same as the databse user. The built_ins 
config have been updated to use this new default DSN.