remove more vestigial web UI support in the login form handler
The `POST .../authn/session` handler is for REST API clients and
should never try to redirect to a referring web page.  We'll return
the success or error status code and stop.

The oauth workflow will involve a callback to the `.../authn/session`
endpoint as a redirect from the external identity provider, and in
this case the preauth_referrer will be set to redirect the client back
to the requesting page, using a different code path not touched by
this commit.