Always return a redirect_url after DELETE /session. This is controlled by configuration parameters:
logout_no_session_path - path to use if there's no valid session

default_logout_path - default path to use if no other appopriate URL is found (i.e., if the logout
succeeds and the provider doesn't supply a URL or if there's no valid session and login_no_session_path
is not set).

This is related to bug #30.