Add new webauthn2-manage addrobot sub-command
Usage example:

    webauthn2-manage addrobot robot1 /home/secrets/robot1_credential.json 90

If successful, this will create a fake robot identity and active
session that is set to expire `90` days in the future, saving the
session credential to `/home/secrets/robot1_credential.json` in the
format expected by deriva-py.

The fake identity will have a trivial structure with the given local
name, e.g. `robot1` used as display\_name and qualified with a local
prefix such as `https://fqdn/webauthn_robot/` when used as the client
identifier. The identity will have no attributes other than its
identifier. The `fqdn` will be the canonical hostname of the local
webauthn server where the session is created.